[hack] pistachio: add development key
Add a hardcoded development key pair so that they can be used to
demonstrate verified boot without having to add it with mkimage
which adds a dependency of the fitimage already existing.

This change cannot be upstreamed. Note also that because this also
distributes the private key no system protected by this can be
condidered secure.

Signed-off-by: Ian Pozella <Ian.Pozella@imgtec.com>