mkimage: fit: check cmd string for buffer overflow
When generating the command to create the fit image the inputs are
file paths which can overflow the buffer. snprintf was being used to
avoid the overflow itself but nothing actually checks if a
truncation occured giving unexpected filenames causing following
stages to fail.

Therefore add some checks before attempting to run the command.

Additionally have bumped up the size of the buffer as the current
inputs can easily get truncated with long paths.

Signed-off-by: Ian Pozella <Ian.Pozella@imgtec.com>