marduk: reset the tpm
The current version of the board has the tpm reset line connected to
a gpio. This causes security issues meaning we need a new HW spin but
to make do with what we have at least reset the tpm on boot.

Therfore have added mfio_setup_tpm function and call it as part of
tpm initialisation.

This is certainly not the best place to put this but considering the
current arch of 'pistachio_bub is king' there isn't much else we can do
without refactoring everything.

Signed-off-by: Ian Pozella <Ian.Pozella@imgtec.com>